Add reusable helper function to find device nodes
